+++ src/Condition/Conditions.php
@@ -5,15 +5,17 @@ namespace FiiSoft\Jackdaw\Condition;
 use FiiSoft\Jackdaw\Condition\Adapter\FilterAdapter;
 use FiiSoft\Jackdaw\Condition\Adapter\PredicateAdapter;
 use FiiSoft\Jackdaw\Filter\Filter;
+use FiiSoft\Jackdaw\Internal\Check;
 use FiiSoft\Jackdaw\Predicate\Predicate;
 
 final class Conditions
 {
     /**
      * @param Condition|Predicate|Filter|callable $condition
+     * @param int $mode
      * @return Condition
      */
-    public static function getAdapter($condition): Condition
+    public static function getAdapter($condition, int $mode = Check::VALUE): Condition
     {
         if ($condition instanceof Condition) {
             return $condition;
@@ -24,11 +26,11 @@ final class Conditions
         }
     
         if ($condition instanceof Filter) {
-            return self::filter($condition);
+            return self::filter($condition, $mode);
         }
     
         if ($condition instanceof Predicate) {
-            return self::predicate($condition);
+            return self::predicate($condition, $mode);
         }
         
         throw new \InvalidArgumentException('Invalid param condition');
@@ -39,14 +41,14 @@ final class Conditions
         return new GenericCondition($condition);
     }
     
-    public static function filter(Filter $filter): FilterAdapter
+    public static function filter(Filter $filter, int $mode = Check::VALUE): FilterAdapter
     {
-        return new FilterAdapter($filter);
+        return new FilterAdapter($filter, $mode);
     }
     
-    public static function predicate(Predicate $predicate): PredicateAdapter
+    public static function predicate(Predicate $predicate, int $mode = Check::VALUE): PredicateAdapter
     {
-        return new PredicateAdapter($predicate);
+        return new PredicateAdapter($predicate, $mode);
     }
     
     /**