+++ src/Discriminator/Adapter/PredicateAdapter.php
@@ -3,15 +3,19 @@
 namespace FiiSoft\Jackdaw\Discriminator\Adapter;
 
 use FiiSoft\Jackdaw\Discriminator\Discriminator;
+use FiiSoft\Jackdaw\Internal\Check;
 use FiiSoft\Jackdaw\Predicate\Predicate;
 
 final class PredicateAdapter implements Discriminator
 {
     private Predicate $predicate;
     
-    public function __construct(Predicate $predicate)
+    private int $mode;
+    
+    public function __construct(Predicate $predicate, int $mode = Check::VALUE)
     {
         $this->predicate = $predicate;
+        $this->mode = Check::getMode($mode);
     }
     
     /**
@@ -19,6 +23,6 @@ final class PredicateAdapter implements Discriminator
      */
     public function classify($value, $key)
     {
-        return $this->predicate->isSatisfiedBy($value, $key);
+        return $this->predicate->isSatisfiedBy($value, $key, $this->mode);
     }
 }