+++ src/Discriminator/ByValue.php
@@ -1,0 +1,14 @@
+<?php declare(strict_types=1);
+
+namespace FiiSoft\Jackdaw\Discriminator;
+
+final class ByValue implements Discriminator
+{
+    /**
+     * @inheritDoc
+     */
+    public function classify($value, $key)
+    {
+        return $value;
+    }
+}