+++ src/Operation/MapWhen.php
@@ -8,6 +8,7 @@ use FiiSoft\Jackdaw\Filter\Filter;
 use FiiSoft\Jackdaw\Internal\Signal;
 use FiiSoft\Jackdaw\Mapper\Mapper;
 use FiiSoft\Jackdaw\Mapper\Mappers;
+use FiiSoft\Jackdaw\Mapper\Value;
 use FiiSoft\Jackdaw\Operation\Internal\BaseOperation;
 use FiiSoft\Jackdaw\Predicate\Predicate;
 use FiiSoft\Jackdaw\Reducer\Reducer;
@@ -31,6 +32,9 @@ final class MapWhen extends BaseOperation
         
         if ($elseMapper !== null) {
             $this->elseMapper = Mappers::getAdapter($elseMapper);
+            if ($this->elseMapper instanceof Value) {
+                $this->elseMapper = null;
+            }
         }
     }
     
@@ -45,5 +49,10 @@ final class MapWhen extends BaseOperation
         }
     
         $this->next->handle($signal);
+    }
+    
+    public function isBarren(): bool
+    {
+        return $this->elseMapper === null && $this->mapper instanceof Value;
     }
 }