+++ src/Reducer/Count.php
@@ -1,0 +1,35 @@
+<?php declare(strict_types=1);
+
+namespace FiiSoft\Jackdaw\Reducer;
+
+use FiiSoft\Jackdaw\Internal\Item;
+
+final class Count implements Reducer
+{
+    private int $count = 0;
+    
+    public function consume($value): void
+    {
+        ++$this->count;
+    }
+    
+    public function result(): int
+    {
+        return $this->count;
+    }
+    
+    public function reset(): void
+    {
+        $this->count = 0;
+    }
+    
+    public function hasResult(): bool
+    {
+        return true;
+    }
+    
+    public function getResult(): Item
+    {
+        return new Item(0, $this->count);
+    }
+}